+++ src/Settings/PreferencesWindow.cpp
@@ -120,7 +120,7 @@ PreferencesWindow::PreferencesWindow(QWidget *parent) : QMainWindow(parent)
 
     addPage("Actions/General", {"Hot Exit/Enable"});
 
-    addPage("Actions/Save", {"Auto Save", "Save Faster", "Auto Format", "Save Tests"});
+    addPage("Actions/Save", {"Auto Save", "Save Faster", "Auto Format", "Save File On Compilation", "Save File On Execution", "Save Tests"});
 
     addPage("Extensions/Clang Format", {"Clang Format/Path", "Clang Format/Style"}, false);